Worked with sessions and authentization when deleting, edit and viewing

+++ App/Controllers/ReportsController.php
@@ -81,24 +81,31 @@ class ReportsController extends Controller {
     }
 
     public function delete($id) {
-        if(!empty($_POST) && Auth::checkCSRF($_POST["token"])) {
-            $model = new ReportsModel();
-            $file  = $model->find($id)->file;
-            unlink(__DIR__ . '/../../public/uploads/' . $file);
-            $model->delete($id);
+        $model = new ReportsModel();
+        $report = $model->find($id);
+        if ($this->isOwner($report)) {
+            if(!empty($_POST) && Auth::checkCSRF($_POST["token"])) {
+                $file  = $model->find($id)->file;
+                unlink(__DIR__ . '/../../public/uploads/' . $file);
+                $model->delete($id);
 
-            App::redirect('reports');
+                App::redirect('reports');
+            }
+
+            else {
+                $model = new ReportsModel();
+                $data = $model->find($id);
+                $this->render('pages/reports_delete.twig', [
+                    'title'       => 'Delete report',
+                    'description' => 'Reports - Just a simple inventory management system.',
+                    'page'        => 'reports',
+                    'data'        => $data
+                ]);
+            }
         }
-
         else {
-            $model = new ReportsModel();
-            $data = $model->find($id);
-            $this->render('pages/reports_delete.twig', [
-                'title'       => 'Delete report',
-                'description' => 'Reports - Just a simple inventory management system.',
-                'page'        => 'reports',
-                'data'        => $data
-            ]);
+            echo "You don't own this report, thereby you can't delete it.";
+            App::error403();
         }
     }
